games/star-catcher/pyproject.toml
[project]
name = "star-catcher"
version = "0.1.0"
description = "Catch falling stars before they hit the ground!"
requires-python = ">=3.10,<4.0"
dependencies = [
    "retro-games>=1.0.0",
]

[project.scripts]
play = "star_catcher.game:main"

[tool.retro]
author = "Chris"
description = "Move your basket left and right to catch falling stars. Each star is worth 10 points. Miss three stars and the game is over!"
result_file = "result.json"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
